backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
chromadb
pypdf
langchain-text-splitters
openai
scikit-learn

